+++ src/backend/twitterApi/useCases.ts
@@ -248,6 +248,64 @@ export class LatestGraphQLTweetUseCase extends GraphQLTweetUseCase {
   }
 }
 
+export class GuestGraphQLTweetUseCase extends GraphQLTweetUseCase {
+  makeEndpoint(): string {
+    const endpoint = new URL('https://twitter.com/i/api/graphql/0hWvDhmW8YQ-S_ib3azIrw/TweetResultByRestId')
+    endpoint.searchParams.append(
+      'variables',
+      JSON.stringify({
+        tweetId: this.tweetId,
+        withCommunity: false,
+        includePromotedContent: false,
+        withVoice: false,
+      })
+    )
+    endpoint.searchParams.append(
+      'features',
+      JSON.stringify({
+        creator_subscriptions_tweet_preview_api_enabled: false,
+        tweetypie_unmention_optimization_enabled: true,
+        responsive_web_edit_tweet_api_enabled: true,
+        graphql_is_translatable_rweb_tweet_is_translatable_enabled: false,
+        view_counts_everywhere_api_enabled: false,
+        longform_notetweets_consumption_enabled: true,
+        responsive_web_twitter_article_tweet_consumption_enabled: false,
+        tweet_awards_web_tipping_enabled: false,
+        freedom_of_speech_not_reach_fetch_enabled: true,
+        standardized_nudges_misinfo: false,
+        tweet_with_visibility_results_prefer_gql_limited_actions_policy_enabled: true,
+        longform_notetweets_rich_text_read_enabled: false,
+        longform_notetweets_inline_media_enabled: false,
+        responsive_web_graphql_exclude_directive_enabled: true,
+        verified_phone_label_enabled: false,
+        responsive_web_media_download_video_enabled: false,
+        responsive_web_graphql_skip_user_profile_image_extensions_enabled: false,
+        responsive_web_graphql_timeline_navigation_enabled: false,
+        responsive_web_enhance_cards_enabled: false,
+      })
+    )
+    endpoint.searchParams.append(
+      'fieldToggles',
+      JSON.stringify({ withArticleRichContentState: false, withAuxiliaryUserLabels: false })
+    )
+    return endpoint.href
+  }
+
+  parseBody(object: any): TweetVO {
+    const tweet = object.data.tweetResult.result.legacy
+    if (!tweet) throw new TweetParsingError('Cannot parse tweet from response.')
+
+    const user = object.data.tweetResult.result.core.user_results.result
+    if (!user) throw new TweetUserParsingError('Cannot parse tweet user from response.')
+
+    return new TweetVO(tweet, {
+      screen_name: user.legacy.screen_name,
+      name: user.legacy.name,
+      rest_id: user.rest_id,
+    })
+  }
+}
+
 export class MediaTweetUseCases {
   constructor(readonly tweetUseCase: ITweetUseCase) {}
 
@@ -286,6 +344,7 @@ export const createAllApiUseCasesByTweetId = (tweetId: string): ITweetUseCase[] 
   new V2TweetUseCase(tweetId),
   new LatestGraphQLTweetUseCase(tweetId),
   new FallbackGraphQLTweetUseCase(tweetId),
+  new GuestGraphQLTweetUseCase(tweetId),
 ]
 
 export const sortUseCasesByVersion =