+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleases]
+
+### Fixed
+
+- Fixed process in guest mode.
+
+
 ## [4.1.10] - 2023-07-24
 
 ### Added